// src/feature/support-application/steps/family-and-financial-info-step/family-and-financial-info-step.types.ts
export enum HousingStatus {
  HOMEOWNER = "HOMEOWNER",
  RENTER = "RENTER",
  LIVING_WITH_FAMILY = "LIVING_WITH_FAMILY",
  HOMELESS = "HOMELESS",
  TEMPORARY_HOUSING = "TEMPORARY_HOUSING",
  OTHER = "OTHER",
}

export enum MaritalStatus {
  SINGLE = "SINGLE",
  MARRIED = "MARRIED",
}

export enum EmploymentStatus {
  EMPLOYED = "EMPLOYED",
  UNEMPLOYED = "UNEMPLOYED",
}